Split validations.rb into different logical components
Now that the design has settled, it is preferable to reduce the
complexity of a file by separating its components into their own smaller
files.
