+++ lib/sheetah/messaging/validations/invalid_message.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require_relative "../../errors/error"
+
+module Sheetah
+  module Messaging
+    module Validations
+      class InvalidMessage < Errors::Error
+      end
+    end
+  end
+end
